requirements.txt
anthropic>=0.38.0
langgraph>=0.2.43
pydantic>=2.7.0
python-dotenv>=1.0.1
rich>=13.7.1
fastapi>=0.112.0
uvicorn[standard]>=0.30.5
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
mmengine>=0.10.7
aiosqlite>=0.20.0

